chore(platform-logs): move under zango.apps.shared (SHARED_APP convention)
`platform_logs` was placed under zango.apps/ but it's registered as a
SHARED_APP. The convention in this codebase is for shared apps to live
under zango.apps.shared/ alongside `tenancy` and `platformauth`.

- Move zango/apps/platform_logs/ → zango/apps/shared/platform_logs/
- Update AppConfig.name and all import sites accordingly
- INSTALLED_APPS string in settings/base.py updated to match
- App label stays `platform_logs` (last segment unchanged), so the
  django_migrations table still recognizes the existing 0001_initial
  record — no DB migration required
