+++ backend/src/zango/apps/shared/platform_logs/apps.py
@@ -3,13 +3,13 @@ from django.apps import AppConfig
 
 class PlatformLogsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "zango.apps.platform_logs"
+    name = "zango.apps.shared.platform_logs"
     verbose_name = "Platform Logs"
 
     def ready(self):
         # Wire Celery's setup_logging signal so worker boot doesn't
         # discard Django's tenant-aware LOGGING dict. Implemented in P4.
         try:
-            from zango.apps.platform_logs import celery_logging  # noqa: F401
+            from zango.apps.shared.platform_logs import celery_logging  # noqa: F401
         except ImportError:
             pass
